# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Restfull CRUD doc",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://localhost/api"
        }
    ],
    "paths": {
        "/api/bem-vindo": {
            "get": {
                "tags": [
                    "/bem-vindo"
                ],
                "summary": "Exibe uma mensagem de boas-vindas",
                "operationId": "84375e9fd6d1310c935a4e8c2c350dc8",
                "responses": {
                    "200": {
                        "description": "Mensagem de boas vindas exibida com sucesso!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Welcome message"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
        "/api/user/get-all": {
            "get": {
                "tags": [
                    "/user"
                ],
                "summary": "Obtém todos os usuários",
                "operationId": "269d2ee527b70c0534b862303e2dd2bf",
                "responses": {
                    "200": {
                        "description": "Busca por todos os usuários realizada com sucesso!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "204": {
                        "description": "Não encontrou usuários"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
        "/api/user/get-active": {
            "get": {
                "tags": [
                    "/user"
                ],
                "summary": "Obtém os usuários ativos",
                "operationId": "8905750e00069511a256ed98d8acde30",
                "responses": {
                    "200": {
                        "description": "Lista de todos os usuários ativos realizada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "204": {
                        "description": "Nenhum usuário ativo encontrado"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
        "/api/user/{id}": {
            "get": {
                "tags": [
                    "/user"
                ],
                "summary": "Obtém um usuário pelo ID",
                "operationId": "7478e3e4aae6f2a53ad2db2858432ccf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do usuário",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuário obtido com sucesso!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "204": {
                        "description": "Usuário não encontrado"
                    },
                    "400": {
                        "description": "Parâmetro ID é obrigatório"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "/user"
                ],
                "summary": "Atualiza um usuário",
                "operationId": "6d61dfebd85e9da159e7f188e49ef43b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserUpdatePayload"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "/user"
                ],
                "summary": "Realiza softDelete em um usuário",
                "operationId": "55b67c0b5a754123ca2f1ff73291b649",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Usuario deletado com sucesso!"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
        "/api/user": {
            "post": {
                "tags": [
                    "/user"
                ],
                "summary": "Cria um novo usuário",
                "operationId": "37f9107a5ce33dbc67845693ce4052d4",
                "requestBody": {
                    "description": "Dados do usuário",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserCreatePayload"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuário criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Usuario criado com sucesso!"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "O email escolhido já possui cadastro"
                    },
                    "500": {
                        "description": "Ocorreu um erro não identificado"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "UserCreatePayload": {
                "required": [
                    "name",
                    "email",
                    "password",
                    "celular",
                    "sexo",
                    "data_nascimento"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "johndoe@example.com"
                    },
                    "password": {
                        "type": "string",
                        "example": "secret123"
                    },
                    "celular": {
                        "type": "string",
                        "example": "(99) 99999-9999"
                    },
                    "sexo": {
                        "type": "string",
                        "enum": [
                            "Masculino",
                            "Feminino"
                        ]
                    },
                    "data_nascimento": {
                        "type": "string",
                        "format": "date",
                        "example": "2000-01-01"
                    },
                    "avatar": {
                        "type": "string",
                        "nullable": true
                    },
                    "active": {
                        "type": "boolean",
                        "default": true
                    }
                },
                "type": "object"
            },
            "UserUpdatePayload": {
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "johndoe@example.com"
                    },
                    "password": {
                        "type": "string",
                        "example": "secret123"
                    },
                    "celular": {
                        "type": "string",
                        "example": "(99) 99999-9999"
                    },
                    "sexo": {
                        "type": "string",
                        "enum": [
                            "Masculino",
                            "Feminino"
                        ]
                    },
                    "data_nascimento": {
                        "type": "string",
                        "format": "date",
                        "example": "2000-01-01"
                    },
                    "avatar": {
                        "type": "string",
                        "nullable": true
                    },
                    "active": {
                        "type": "boolean"
                    }
                },
                "type": "object"
            }
        }
    }
}